Move SMT operations
